Add an --extensions switch to run_test (#964)
When passing that switch the install and uninstall of spatial db
happens using the EXTENSION model. It's currently not run as part
of "make check" because some tests fail due to use of ambiguos
calls when both core and raster modules are installed (there's no
way to separate them in the EXTENSION model).

git-svn-id: http://svn.osgeo.org/postgis/trunk@9243 b70326c6-7e19-0410-871a-916f4a2858ee