TopoGeo_addPoint: use a more functional tolerance when snapping (#1613)
All of this looks like magic but it isn't.
I actually wonder if ST_ModEdgeSplit and ST_ModEdgesSplit and ST_Split
itself should do this internally, and if in doing so we wouldn't need
to do it from higher levels. It doesn't indeed feel comfortable to do
all this noise on such an high level.

Anyway this commit adds a now-passing regression test for the topology
building issue and that's A Good Thing.

git-svn-id: http://svn.osgeo.org/postgis/trunk@9330 b70326c6-7e19-0410-871a-916f4a2858ee